movendo biblioteca para a pasta funcionalidades
